extract/requirements.txt
black==22.12.0
flake8==6.0.0
google-api-core>=2.1.1
google-auth>=2.3.0
google-cloud-core>=2.1.0
google-cloud-storage>=1.42.3
pytest==7.2.0